+++ src/main/java/io/gravitee/resource/oauth2/am/OAuth2AMResource.java
@@ -20,24 +20,33 @@ import io.gravitee.common.http.HttpStatusCode;
 import io.gravitee.common.http.MediaType;
 import io.gravitee.common.utils.UUID;
 import io.gravitee.gateway.api.handler.Handler;
+import io.gravitee.gateway.reactive.api.context.DeploymentContext;
 import io.gravitee.node.api.Node;
 import io.gravitee.node.api.configuration.Configuration;
 import io.gravitee.node.api.utils.NodeUtils;
 import io.gravitee.node.container.spring.SpringEnvironmentConfiguration;
+import io.gravitee.node.vertx.client.http.VertxHttpClientFactory;
 import io.gravitee.node.vertx.proxy.VertxProxyOptionsUtils;
+import io.gravitee.plugin.mappers.HttpClientOptionsMapper;
+import io.gravitee.plugin.mappers.HttpProxyOptionsMapper;
+import io.gravitee.plugin.mappers.SslOptionsMapper;
 import io.gravitee.resource.oauth2.am.configuration.OAuth2ResourceConfiguration;
+import io.gravitee.resource.oauth2.am.configuration.OAuth2ResourceConfigurationEvaluator;
 import io.gravitee.resource.oauth2.api.OAuth2Resource;
 import io.gravitee.resource.oauth2.api.OAuth2ResourceException;
+import io.gravitee.resource.oauth2.api.OAuth2ResourceMetadata;
 import io.gravitee.resource.oauth2.api.OAuth2Response;
 import io.gravitee.resource.oauth2.api.openid.UserInfoResponse;
 import io.vertx.core.AsyncResult;
-import io.vertx.core.Vertx;
 import io.vertx.core.http.*;
 import io.vertx.core.json.JsonObject;
+import io.vertx.rxjava3.core.Vertx;
+import java.net.URI;
 import java.net.URL;
 import java.util.Base64;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
+import java.util.List;
+import javax.inject.Inject;
+import lombok.Setter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.BeansException;
@@ -69,20 +78,32 @@ public class OAuth2AMResource extends OAuth2Resource<OAuth2ResourceConfiguration
     private static final String PATH_SEPARATOR = "/";
     private ApplicationContext applicationContext;
 
-    private final Map<Thread, HttpClient> httpClients = new ConcurrentHashMap<>();
+    private HttpClient httpClient;
 
-    private HttpClientOptions httpClientOptions;
-
-    private Vertx vertx;
     private String userAgent;
 
     private String introspectionEndpointURI;
     private String introspectionEndpointAuthorization;
     private String userInfoEndpointURI;
+    private OAuth2ResourceConfiguration configuration;
+
+    @Inject
+    @Setter
+    private DeploymentContext deploymentContext;
+
+    @Override
+    public OAuth2ResourceConfiguration configuration() {
+        if (configuration == null) {
+            return super.configuration();
+        }
+        return configuration;
+    }
 
     @Override
     protected void doStart() throws Exception {
         super.doStart();
+
+        configuration = new OAuth2ResourceConfigurationEvaluator(configuration()).evalNow(deploymentContext);
 
         logger.info("Starting an OAuth2 resource using Gravitee.io Access Management server at {}", configuration().getServerURL());
 
@@ -95,32 +116,20 @@ public class OAuth2AMResource extends OAuth2Resource<OAuth2ResourceConfiguration
         // URI.getHost does not support '_' in the name, so we are using an intermediate URL to get the final host
         String authorizationServerHost = introspectionUrl.getHost();
 
-        httpClientOptions =
-            new HttpClientOptions()
-                .setDefaultPort(authorizationServerPort)
-                .setDefaultHost(authorizationServerHost)
-                .setMaxPoolSize(configuration().getHttpClientOptions().getMaxConcurrentConnections())
-                .setIdleTimeout(60)
-                .setConnectTimeout(10000);
+        var target = new URL(introspectionUrl.getProtocol(), authorizationServerHost, authorizationServerPort, introspectionUrl.getFile());
 
-        if (configuration().isUseSystemProxy()) {
-            try {
-                Configuration nodeConfig = new SpringEnvironmentConfiguration(applicationContext.getEnvironment());
-                httpClientOptions.setProxyOptions(VertxProxyOptionsUtils.buildProxyOptions(nodeConfig));
-            } catch (IllegalStateException e) {
-                logger.warn(
-                    "AMResource requires a system proxy to be defined to call [{}] but some configurations are missing or not well defined: {}",
-                    configuration().getServerURL(),
-                    e.getMessage()
-                );
-                logger.warn("Ignoring system proxy");
-            }
-        }
-
-        // Use SSL connection if authorization schema is set to HTTPS
-        if (HTTPS_SCHEME.equalsIgnoreCase(introspectionUrl.getProtocol())) {
-            httpClientOptions.setSsl(true).setVerifyHost(false).setTrustAll(true);
-        }
+        httpClient =
+            VertxHttpClientFactory
+                .builder()
+                .vertx(applicationContext.getBean(Vertx.class))
+                .nodeConfiguration(new SpringEnvironmentConfiguration(applicationContext.getEnvironment()))
+                .defaultTarget(target.toString())
+                .httpOptions(HttpClientOptionsMapper.INSTANCE.map(configuration().getHttpClientOptions()))
+                .sslOptions(SslOptionsMapper.INSTANCE.map(configuration().getSslOptions()))
+                .proxyOptions(HttpProxyOptionsMapper.INSTANCE.map(configuration().getHttpProxyOptions()))
+                .build()
+                .createHttpClient()
+                .getDelegate();
 
         introspectionEndpointAuthorization =
             AUTHORIZATION_HEADER_BASIC_SCHEME +
@@ -149,28 +158,20 @@ public class OAuth2AMResource extends OAuth2Resource<OAuth2ResourceConfiguration
         }
 
         userAgent = NodeUtils.userAgent(applicationContext.getBean(Node.class));
-        vertx = applicationContext.getBean(Vertx.class);
     }
 
     @Override
     protected void doStop() throws Exception {
         super.doStop();
-
-        httpClients
-            .values()
-            .forEach(httpClient -> {
-                try {
-                    httpClient.close();
-                } catch (IllegalStateException ise) {
-                    logger.warn(ise.getMessage());
-                }
-            });
+        try {
+            httpClient.close();
+        } catch (IllegalStateException ise) {
+            logger.warn(ise.getMessage());
+        }
     }
 
     @Override
     public void introspect(String accessToken, Handler<OAuth2Response> responseHandler) {
-        HttpClient httpClient = httpClients.computeIfAbsent(Thread.currentThread(), context -> vertx.createHttpClient(httpClientOptions));
-
         logger.debug("Introspect access token by requesting {}", introspectionEndpointURI);
 
         final RequestOptions reqOptions = new RequestOptions()
@@ -263,8 +264,6 @@ public class OAuth2AMResource extends OAuth2Resource<OAuth2ResourceConfiguration
 
     @Override
     public void userInfo(String accessToken, Handler<UserInfoResponse> responseHandler) {
-        HttpClient httpClient = httpClients.computeIfAbsent(Thread.currentThread(), context -> vertx.createHttpClient(httpClientOptions));
-
         logger.debug("Get userinfo from {}", userInfoEndpointURI);
 
         final RequestOptions reqOptions = new RequestOptions()
@@ -350,5 +349,12 @@ public class OAuth2AMResource extends OAuth2Resource<OAuth2ResourceConfiguration
     @Override
     public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
         this.applicationContext = applicationContext;
+    }
+
+    @Override
+    public OAuth2ResourceMetadata getProtectedResourceMetadata(String protectedResourceUri) {
+        URI authServerUri = URI.create(configuration().getServerURL() + "/" + configuration().getSecurityDomain() + "/oidc");
+        String authorizationServer = authServerUri.normalize().toString().replaceAll("/+$", "");
+        return new OAuth2ResourceMetadata(protectedResourceUri, List.of(authorizationServer), List.of());
     }
 }